+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 0.9.5 &ndash; 2023-06-12
+
 ### Added
 
 * Call Python methods from a strategy (#362)